Delivery time slots, phone validation, refactoring & UI improvements

+++ backend/src/backend/presentation/http/v1/schemas/product.py
@@ -1,7 +1,9 @@
+from decimal import Decimal
+
 from pydantic import BaseModel
 
 
 class EditProductSchema(BaseModel):
     name: str | None = None
-    price: int | None = None
+    price: Decimal | None = None
     category_id: str | None = None
